src_dilluns_18/requirements.txt
python-chess
pandas
cairosvg